New, slightly less jank build system working
Still need to get the debug flags working, but we have a slightly
improved build system for the kernel and bootstrap programs. Creating
the ISO image is now designated to bootstrap. Both the kernel and
bootstrap recipes now use raw makefiles; autotools is a bit of a hassel
that I do not want to poke.