Implementing external ports
I have created a brand new external folder which will contain all of the
Makefiles to prepare the initramfs. It's definitely a bit hacky, but so
far mlibc and bash-5.2 manage to successfully compile and install into
the initramfs. Next I need to port gcc and some other dependencies of bash.