Successful toolchain-host compilation
After some time, I have finally managed to build the host-toolchain. It
has all worked successfully and the correct files seem to be properly
installed into the sysroot. There most definitely needs to be
improvements with everything. I am quite suspicious of the host AC_FLAGS
including --with-sysroot=$(ARC_SYSROOT) instead of--with-sysroot=/
--with-build-sysroot=$(ARC_SYSROOT).