Update mlibc patch
There were a lot of issues as the compiler was optimizing out way too much. It inlined functions without understanding the purpose of them (in the libc panic function it would pass improper parameters to the libc log syscall and then return early, not calling the exit function, and waiting defeating the purpose of the panic function all together). Sometimes the arguments would not be arranged or modified as the compiler assumed it had full control. Overall, I have just turned off optimization for the syscalls, though now it does preform a lot of extra steps which are not needed.