Fix autreconf.sh, no need for O0
I have fixed the recursive_autoreconf script. It was previously giving errors because I resolved a variable as if it were a Makefile instead of a shell script. The mlibc patch has also been updated and no longer needs the disabling of optimization. Each function which calls a syscall can be made into a naked function causing the compiler to forgo assumptions about argument order and actually placing things into the correct registers.