Add tar_helper.sh
I have added a script to aid with the downloading of TAR files from online mirrors. I have also went through and added in some standard variables for the various package build scripts in toolchain-host and toolchain-build to interact with this script. Since the variable ARC_SOURCE_DIR was introduced, there is no more need for the ARC_RECONF_DIR in the auto reconfigurator script. I have not tested this to see if it works, I am making this as a back up of sorts.