config.h: changes
The ARC_SEPARATE_LOW_MEM option has been removed as that is now
specified in the PMM bias elements. The process file limit has been
downgraded from 1024 to 512, it could also be pushed higher. This was
done such that less memory is wasted by the processor descriptor.
Further, two additional values have been added to control the default
size of stack that is allocated for kernel space use and userspace use.